requirements.txt

langchain-google-genai
langchain-community==0.2.12
langchain==0.2.14
pymilvus
langchain-groq
sentence-transformers

pypdf
python-dotenv
tiktoken
playwright
fastapi
uvicorn
python-multipart
